# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TrackType = (typeof TrackType)[keyof typeof TrackType];

export const TrackType = {
  original: "original",
  remix: "remix",
  live: "live",
  cover: "cover",
} as const;

export type TrackSource = (typeof TrackSource)[keyof typeof TrackSource];

export const TrackSource = {
  youtube: "youtube",
  soundcloud: "soundcloud",
  bandcamp: "bandcamp",
  deezer: "deezer",
} as const;

export interface TrackResult {
  /** Unique track ID. Format: yt_<base64url(https://www.youtube.com/watch?v=ID)> for YouTube or sc_<base64url(permalink_url)> for SoundCloud. Pass this ID to /stream or /download endpoints. */
  id: string;
  title: string;
  artist: string;
  type: TrackType;
  /** Duration in seconds */
  duration: number;
  source: TrackSource;
  thumbnailUrl?: string | null;
  /** Available quality options (e.g. ["128", "320"]) */
  quality: string[];
  viewCount?: number | null;
  /** Ranking score (higher is better match) */
  score: number;
}

export type SearchRequestMode =
  (typeof SearchRequestMode)[keyof typeof SearchRequestMode];

export const SearchRequestMode = {
  auto: "auto",
  manual: "manual",
} as const;

export type SearchRequestSourcesItem =
  (typeof SearchRequestSourcesItem)[keyof typeof SearchRequestSourcesItem];

export const SearchRequestSourcesItem = {
  yt: "yt",
  sc: "sc",
  bc: "bc",
  dz: "dz",
} as const;

export interface SearchRequest {
  /**
   * Artist name, trimmed before validation
   * @minLength 1
   * @maxLength 200
   */
  artist: string;
  /**
   * Track title, trimmed before validation
   * @minLength 1
   * @maxLength 300
   */
  title: string;
  mode?: SearchRequestMode;
  /**
   * @minItems 1
   * @maxItems 4
   */
  sources?: SearchRequestSourcesItem[];
  /**
   * @minimum 1
   * @maximum 40
   */
  maxResults?: number;
}

export type SearchResponseSourcesItem =
  (typeof SearchResponseSourcesItem)[keyof typeof SearchResponseSourcesItem];

export const SearchResponseSourcesItem = {
  yt: "yt",
  sc: "sc",
  bc: "bc",
  dz: "dz",
} as const;

export interface SearchResponse {
  /** The normalized search query used */
  query: string;
  results: TrackResult[];
  /** Whether results came from cache */
  cached: boolean;
  /**
   * @minItems 1
   * @maxItems 4
   */
  sources: SearchResponseSourcesItem[];
  /** Whether broadening a manual source selection may return results */
  fallbackAvailable: boolean;
}

export interface StreamResponse {
  id: string;
  streamUrl: string;
  mimeType?: string | null;
}

export interface DownloadResponse {
  id: string;
  downloadUrl: string;
  filename?: string | null;
}

export type DownloadQuality =
  (typeof DownloadQuality)[keyof typeof DownloadQuality];

export const DownloadQuality = {
  NUMBER_128: "128",
  NUMBER_192: "192",
  NUMBER_256: "256",
  NUMBER_320: "320",
  flac: "flac",
} as const;

export interface DownloadQueueTrack {
  /**
   * @minLength 1
   * @maxLength 4096
   */
  trackId: string;
  /**
   * @minLength 1
   * @maxLength 300
   */
  artist: string;
  /**
   * @minLength 1
   * @maxLength 500
   */
  title: string;
  quality: DownloadQuality;
}

export interface DownloadQueueRequest {
  /**
   * @minItems 1
   * @maxItems 50
   */
  tracks: DownloadQueueTrack[];
}

export interface DownloadQueueSuccessResult {
  /**
   * @minLength 1
   * @maxLength 4096
   */
  trackId: string;
  jobId: string;
  /**
   * @minimum 1
   * @maximum 200
   */
  position: number;
}

export interface DownloadQueueFailureResult {
  /**
   * @minLength 1
   * @maxLength 4096
   */
  trackId: string;
  error: "download_queue_unavailable";
}

export type DownloadQueueResult =
  | DownloadQueueSuccessResult
  | DownloadQueueFailureResult;

export interface DownloadQueueResponse {
  /**
   * @minItems 1
   * @maxItems 50
   */
  results: DownloadQueueResult[];
}

export type DownloadJobState =
  (typeof DownloadJobState)[keyof typeof DownloadJobState];

export const DownloadJobState = {
  waiting: "waiting",
  active: "active",
  completed: "completed",
  failed: "failed",
  canceled: "canceled",
} as const;

export interface DownloadJobStatus {
  status: DownloadJobState;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  /**
   * @minimum 1
   * @maximum 200
   */
  position?: number;
  /**
   * @minimum 1
   * @maximum 1073741824
   */
  fileSize?: number;
}

export interface DownloadJob {
  jobId: string;
  status: DownloadJobState;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  /**
   * @minimum 1
   * @maximum 200
   */
  position?: number;
  /**
   * @minimum 1
   * @maximum 1073741824
   */
  fileSize?: number;
}

export interface DownloadJobListResponse {
  /** @maxItems 200 */
  jobs: DownloadJob[];
}

export interface DownloadJobCancelResponse {
  jobId: string;
  status: DownloadJobState;
}

export type DownloadErrorResponseError =
  (typeof DownloadErrorResponseError)[keyof typeof DownloadErrorResponseError];

export const DownloadErrorResponseError = {
  bad_request: "bad_request",
  module_access_denied: "module_access_denied",
  download_queue_unavailable: "download_queue_unavailable",
  job_not_found: "job_not_found",
  file_not_found: "file_not_found",
  file_not_ready: "file_not_ready",
  range_not_satisfiable: "range_not_satisfiable",
  worker_unavailable: "worker_unavailable",
} as const;

export interface DownloadErrorResponse {
  error: DownloadErrorResponseError;
}

export interface LikedTrack {
  /**
   * @minLength 4
   * @maxLength 4096
   * @pattern ^(yt|sc|bc|dz)_[A-Za-z0-9_-]+$
   */
  trackId: string;
  /** @maxLength 300 */
  artist: string | null;
  /** @maxLength 500 */
  title: string | null;
  /** @maxLength 2048 */
  thumbnailUrl: string | null;
  /**
   * @minimum 1
   * @maximum 86400
   */
  durationSeconds: number | null;
  likedAt: string;
}

export interface LikedTrackPage {
  /** @maxItems 100 */
  items: LikedTrack[];
  /**
   * @maxLength 64
   * @pattern ^[A-Za-z0-9_-]+$
   */
  nextCursor: string | null;
}

export interface SaveLikedTrackRequest {
  /**
   * @minLength 1
   * @maxLength 300
   */
  artist: string;
  /**
   * @minLength 1
   * @maxLength 500
   */
  title: string;
  /** @maxLength 2048 */
  thumbnailUrl?: string | null;
  /**
   * @minimum 1
   * @maximum 86400
   */
  durationSeconds?: number | null;
}

export interface LikedTrackMutationResponse {
  item: LikedTrack;
}

export type CollectionErrorResponseError =
  (typeof CollectionErrorResponseError)[keyof typeof CollectionErrorResponseError];

export const CollectionErrorResponseError = {
  bad_request: "bad_request",
  unauthorized: "unauthorized",
  forbidden: "forbidden",
  module_access_denied: "module_access_denied",
  policy_unavailable: "policy_unavailable",
  internal_error: "internal_error",
} as const;

export interface CollectionErrorResponse {
  error: CollectionErrorResponseError;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export const SearchUnavailableResponseValue = {
  error: "search_unavailable",
} as const;
export type SearchUnavailableResponse = typeof SearchUnavailableResponseValue;

export type ListLikedTracksParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @minLength 1
   * @maxLength 64
   * @pattern ^[A-Za-z0-9_-]+$
   */
  cursor?: string;
};