# lib/api-zod/src/generated/types/likedTrackPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Player API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LikedTrack } from "./likedTrack";

export interface LikedTrackPage {
  /** @maxItems 100 */
  items: LikedTrack[];
  /**
   * @maxLength 64
   * @pattern ^[A-Za-z0-9_-]+$
   */
  nextCursor: string | null;
}